ml-service/requirements.txt
# FinTrack Pro ML Service Dependencies

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ML/Data Science
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
statsmodels==0.14.0

# Database
pymongo==4.6.1
motor==3.3.2
redis==5.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3

# Logging
loguru==0.7.2

# OCR / Image Processing
pytesseract==0.3.10
Pillow==10.1.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
